added time domain matrix iteration.
It allows to make infinity experiments once we have the matrix, but the
matrix is very expensive to compute.
There is a problem of nonlinearity of the convergence.
The figure fig_*_c_zero illustrate it.
There is a problem in the low frequencies,
and there is an other problem (of a similar nature ?) when taking big h.